重构诗文模块

+++ app/src/main/java/com/hefengbao/jingmo/data/database/entity/ChineseWisecrackEntity.kt
@@ -10,7 +10,7 @@ import androidx.room.PrimaryKey
 @Entity("chinese_wisecracks")
 data class ChineseWisecrackEntity(
     @PrimaryKey
-    val id: Long,
+    val id: Int,
     val riddle: String,
     val answer: String,
     @ColumnInfo(name = "first_word")
